requirements.txt
configparser==3.7.4
python-dateutil==2.8.0
requests==2.19.1
urllib3==1.23
